+++ week4/KangMinhee/src/components/FloatingButton.jsx
@@ -1,0 +1,12 @@
+
+import "../App.css";
+
+const FloatingButton = () => {
+  return (
+    <button className="floating-btn">
+      <img src="/plus.svg" alt="plus" className="plus-icon" />
+    </button>
+  );
+};
+
+export default FloatingButton;
